MAIN MODULE, CUSTOMER DASHBOARD, LOGIN, REGISTER
The primary goal of this banking application is to provide a user-friendly and secure platform for customers to manage their accounts and conduct banking transactions seamlessly. By adhering to a layered architecture approach, the application ensures a clear separation of concerns, making it scalable, maintainable, and extensible for potential future endeveours.